requirements.txt
Django==3.2.5
psycopg2-binary>=2.9.6
djangorestframework==3.12.4
gunicorn==20.1.0
bcrypt==4.2.0
PyJWT==2.8.0
django-environ==0.4.5
djangorestframework-simplejwt
